EnumTable now known as ChoiceTable
ChoiceTable is essentially what EnumTable was, only that it has been generalized to accept all types, not just enums. As a consequence, I imagine the performance of ChoiceTable has dropped, but since we have no performance-critical applications of it, getting ChoiceTable back shouldn't be too hard.

I'll make a ticket to create EnumTable again as a subclass of ChoiceTable

git-svn-id: http://bluespot.unfuddle.com/svn/bluespot_library/trunk@8 93453b9e-65dd-4715-a0fb-a9eb6d1ddc67