Added com.bluespot.logic. This is a collection of two main hierarchies of classes: Adapters and Predicates.
Predicates represent boolean expressions, formed as object. They're comparable to Hamcrest operators, though I think ours are better-done. They're inherently designed to be composites and serve that purpose very well and very cleanly.

One problem with predicates is that they're composed of many tiny objects. This makes their syntax awkward. Another problem is that their immutability means they cannot be constructed on-the-run. Each must be completely built first. To solve both of these problems, we provide predicate builders. These builders allow the illusion of mutabilty, combined with the readable terseness that accompanies the builder pattern. Even more, we provide Predicates, a collection of static methods that create our predicates. This collection can be statically imported for an unprecedented level of readability.

Adapters represent arbitrary, generic converters. The combination of these two concepts enables a surprising rich syntax. We've already begun to use adapters in our code, and I definitely believe we'll end up using them almost everywhere.

The last significant part here is the Visitor/Sentinel interfaces/classes. A visitor is essentially a generic receiver of some data. Sentinels combine the visitor pattern with predicates. Any value passed to the sentinel must be validated with the predicate. If it succeeds, then the sentinel passes that value to its child visitors.

Since Sentinel implements Visitor, it can be chained with other Sentinels. This allows behavior to be expressed in a very clean, declarative way.

git-svn-id: http://bluespot.unfuddle.com/svn/bluespot_library/trunk@104 93453b9e-65dd-4715-a0fb-a9eb6d1ddc67