Application now sets the lookAndFeel before creating the Application. This helps avoid some inconsistent look and feels. Specifically, if components are created in the constructor or in initializers, they will use the default look and feel. This bug didn't crop up in startup(), but it seemed worth fixing.
One complication from this is that we're forced to create an application context first and pass it into the constructor. A default constructor that recreates the context exists, but I'm not sure how stable it is since that would mean two would exist. This should be fixed, and I'm writing a ticket up about it.

git-svn-id: http://bluespot.unfuddle.com/svn/bluespot_library/trunk@232 93453b9e-65dd-4715-a0fb-a9eb6d1ddc67