Changed the way BufferedValue works. It now does not allow a non-null BufferedValue#get(). This allows it to be much more robust, at the expense of visibility.
Since null represents a failed get in most cases, this is actually a totally good thing. If it turns out we need a BufferedValue that is not so judgmental, we can factor one out.

Another direction is a Value that updates a value based on some arbitrary predicate. If it turns out that's a common idiom, we can do it. I actually sort of like this idea, but currently can't think of a use-case. The more general case is to use a Sentinel with a MutableValueVisitor implementation, which seems more practical.

git-svn-id: http://bluespot.unfuddle.com/svn/bluespot_library/trunk@374 93453b9e-65dd-4715-a0fb-a9eb6d1ddc67