Lots more typename cleanup
Generally speaking, I like typenames to stay as close to the underlying
class as possible. It gets a bit difficult when you have a type that's
being adapted and -also- passed to a function, but for now I just defer
to the function's typename. This seems better since this describes the
purpose of the typename, rather than the intermediate step.