Added Adapting- and WrappingCurryable
These were supposed to be the tools that produced SafeCurryable objects,
but it turned out that combining them to create a SafeCurryable was
pretty difficult. It was easier to write AdaptingSafeCurryable which
does both adaptations in one class.

I might remove these later, but they seem mildly useful right now.
They're also already written and relatively bulletproof, so I'm going to
keep them.